(feature) Conference notes in own section
This was fiddly as hell.

The idea is to take all posts which are notes from conference talks, and
separate them from the main blog content.

Jekyll doesn't allow hierarchical categories (they tried, but it broke
too much stuff, and in any case the paginator object doesn't seem to
support filtering by category type, so 'collections' seems to be the way
to go.

Only the first level of folders is actually recognised
(_conference-notes) - the second layer is used in the URL, but only
because it's considered part of the filename: we can't use that to
filter on - i.e. collections aren't hierarchical either.

I could have resolved that by putting a category on each of the posts
corresponding to which conference they're from, but I like having them
in directories, so the approach of "get _everything_ and only display
the ones matching part of the path", though slow and clunky, is fine
because I have few enough posts that builds won't take too long.

Would have been nice to somehow introspect on the directory name to get
the section titles with the conference names, but since I want to add
extra metadata, it seems not too bad to use the _data file to determine
the ordering as well.