Replace Octopress codeblocks with Jekyll highlight
Octopress is basically a dead project. It's a shame: they had nice
codeblocks.

Jekyll doesn't support including the filename so let's just add it as a
comment in the file.

Note that this breaks things a bit - code blocks are indented and don't
have the desired highlighting. I'll fix that in the next commits.