Replace Octopress solarized
...with a slightly hacked together theme.

Hacked together inspired by the specific styles and highlighting choices
of https://github.com/octopress/solarized, but without having to pull in
so many (unmaintained!) dependencies.

Also inspired by the approach of https://github.com/rphlmr/jekyll-solarized-theme

Need to be careful to override some of the styles added by Pixyll (some
of which are very odd - like increasing the size of punctuation).

Note that this was the only reason the asset tag was included - this was
part of Octopress' asset pipeline.