Replace Octopress with jekyll-compose
...and a github action for deployment.

Octopress was exciting to start with, but Jekyll caught up with some of
the features it delivered, and it felt like it was kind of overreaching:
trying to be a WordPress or a Rails. It hasn't had any maintenance for
years.

Jekyll-compose also doesn't look particularly maintained, but at least
it has no runtime dependencies other than Jekyll itself.