Refine syntax highlighting colours
This is a bit challenging to get React and Rails code to look good. The
Ruby highlighting isn't great: it tags both function definitions and
function calls with the same class, and I don't really want to highlight
both in the same way. It also doesn't know about Rails, so doesn't tag
Rails things as keywords.